+++ resources/views/layouts/main-layout.blade.php
@@ -80,6 +80,20 @@
             </div>
         @endif
 
+        <!-- Warning Toast -->
+        @if (session('warning'))
+            <div class="toast align-items-center text-bg-warning border-0 show mb-2" role="alert" aria-live="assertive"
+                 aria-atomic="true">
+                <div class="d-flex">
+                    <div class="toast-body">
+                        {{ session('warning') }}
+                    </div>
+                    <button type="button" class="btn-close me-2 m-auto" data-bs-dismiss="toast"
+                            aria-label="Close"></button>
+                </div>
+            </div>
+        @endif
+
         <!-- Error Toast -->
         @if (session('error'))
             <div class="toast align-items-center text-bg-danger border-0 show" role="alert" aria-live="assertive"